+++ src/store/index.js
@@ -2,6 +2,7 @@ import Vue from 'vue'
 import Vuex from 'vuex'
 import app from './modules/app'
 import errorLog from './modules/errorLog'
+import moderationLog from './modules/moderationLog'
 import invites from './modules/invites'
 import permission from './modules/permission'
 import reports from './modules/reports'
@@ -19,6 +20,7 @@ const store = new Vuex.Store({
   modules: {
     app,
     errorLog,
+    moderationLog,
     invites,
     permission,
     reports,